UOE-8168: Code review changes
