+++ adapters/pubmatic/pubmatictest/supplemental/gptSlotNameInImpExt.json
@@ -38,7 +38,6 @@
             }
           },
           "data": {
-            "pbadslot": "/2222/home",
             "adserver": {
               "name": "gam",
               "adslot": "/1111/home"
